requirements.txt
streamlit==1.23.1
transformers==4.32.0
torch==2.1.0
Pillow==9.5.0
requests==2.31.0
matplotlib==3.8.0

